+++ frontend/techpick/src/app/style.css.ts
@@ -4,3 +4,19 @@ export const rootLayout = style({
   width: '100%',
   height: '100vh',
 });
+
+import { themeContract } from '@/shared/themes';
+
+export const viewWrapper = style({
+  width: '100%',
+  height: '100%',
+  backgroundColor: themeContract.color.background,
+  color: themeContract.color.font,
+});
+
+export const viewContainer = style({
+  width: '100%',
+  height: '100%',
+  display: 'flex',
+  borderRadius: '8px',
+});
