+++ frontend/techpick/src/widgets/DirectoryTreeSection/directoryTreeSection.css.ts
