+++ frontend/techpick/src/widgets/FeaturedSection/FeaturedSection.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { rightSection } from './featuredSection.css';
+
+export function FeaturedSection() {
+  return <div className={rightSection}></div>;
+}
