+++ frontend/techpick/src/widgets/FeaturedSection/featuredSection.css.ts
