+++ frontend/techpick/src/widgets/MiddleSection/linkEditorSection.css.ts
