--- frontend/techpick/src/widgets/RightSection/RightSection.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { rightSection } from './rightSection.css';
-
-export function RightSection() {
-  return <div className={rightSection}></div>;
-}
